src/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
gunicorn==21.2.0
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==23.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
